+++ types/redis/index.d.ts
@@ -15,7 +15,7 @@
 //                 Thomas de Barochez <https://github.com/tdebarochez>
 //                 David Stephens <https://github.com/dwrss>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-
+ 
 // Imported from: https://github.com/types/npm-redis
 
 /// <reference types="node" />
@@ -30,7 +30,7 @@ export interface RetryStrategyOptions {
     attempt: number;
 }
 
-export type RetryStrategy = (options: RetryStrategyOptions) => number | Error | undefined;
+export type RetryStrategy = (options: RetryStrategyOptions) => number | Error | unknown;
 
 export interface ClientOpts {
     host?: string;
@@ -105,6 +105,7 @@ export interface OverloadedSetCommand<T, U, R> {
     (key: string, arg1: T, arg2: T, cb?: Callback<U>): R;
     (key: string, arg1: T | { [key: string]: T } | T[], cb?: Callback<U>): R;
     (key: string, ...args: Array<T | Callback<U>>): R;
+    (args: [string, ...T[]], cb?: Callback<U>): R;
 }
 
 export interface OverloadedLastCommand<T1, T2, U, R> {
@@ -482,14 +483,14 @@ export interface Commands<R> {
     /**
      * Get the value of a hash field.
      */
-    hget(key: string, field: string, cb?: Callback<string | null>): R;
-    HGET(key: string, field: string, cb?: Callback<string | null>): R;
+    hget(key: string, field: string, cb?: Callback<string>): R;
+    HGET(key: string, field: string, cb?: Callback<string>): R;
 
     /**
      * Get all fields and values in a hash.
      */
-    hgetall(key: string, cb?: Callback<{ [key: string]: string } | null>): R;
-    HGETALL(key: string, cb?: Callback<{ [key: string]: string } | null>): R;
+    hgetall(key: string, cb?: Callback<{ [key: string]: string }>): R;
+    HGETALL(key: string, cb?: Callback<{ [key: string]: string }>): R;
 
     /**
      * Increment the integer value of a hash field by the given number.
@@ -1257,15 +1258,17 @@ export function createClient(options?: ClientOpts): RedisClient;
 
 export function print(err: Error | null, reply: any): void;
 
-export class RedisError extends Error { }
+export class RedisError extends Error {
+    name: string;
+}
 export class ReplyError extends RedisError {
     command: string;
-    args?: any[];
+    args?: unknown[];
     code: string;
 }
 export class AbortError extends RedisError {
     command: string;
-    args?: any[];
+    args?: unknown[];
     code?: string;
 }
 export class ParserError extends RedisError {
