+++ types/redis/redis-tests.ts
@@ -8,7 +8,8 @@ const err: Error = new Error();
 const args: any[] = [];
 const resCallback: (err: Error | null, res: any) => void = () => {};
 const numCallback: (err: Error | null, res: number) => void = () => {};
-const strCallback: (err: Error | null, res: string | null) => void = () => {};
+const strCallback: (err: Error | null, res: string) => void = () => {};
+const nullableStrCallback: (err: Error | null, res: string | null) => void = () => {};
 const okCallback: (err: Error | null, res: 'OK') => void = () => {};
 const messageHandler: (channel: string, message: any) => void = () => {};
 
@@ -74,7 +75,7 @@ client.append(str, str, numCallback);
 client.bitcount(str, numCallback);
 client.bitcount(str, num, num, numCallback);
 client.set(str, str, okCallback);
-client.get(str, strCallback);
+client.get(str, nullableStrCallback);
 client.exists(str, numCallback);
 
 // Event handlers
@@ -109,11 +110,15 @@ client.hgetall(str, resCallback);
 // Deprecated commands
 client.hmset(str, value, okCallback);
 client.hmset(str, str, str, str, str, okCallback);
+client.hmset([str, str, str, str, str], okCallback);
+client.hmset([str, str, str, str, str]);
 client.hmset(str, [str, str, str, str]);
 client.hmset(str, [str, value, str, value], okCallback);
 // Redis 4 variadic HSET
 client.hset(str, value, numCallback);
 client.hset(str, str, str, str, str, numCallback);
+client.hset([str, str, str, str, str], numCallback);
+client.hset([str, str, str, str, str]);
 client.hset(str, [str, str, str, str]);
 client.hset(str, [str, value, str, value], numCallback);
